maintaining brushed state
