+++ src/init.js
@@ -1,3 +1,4 @@
+import { set } from 'd3';
 import { multiply } from 'webcharts';
 
 export default function init(data) {
@@ -15,5 +16,28 @@ export default function init(data) {
             return diff ? diff : aPos > -1 ? -1 : bPos > -1 ? 1 : leftSort ? -1 : rightSort ? 1 : 0;
         } else return leftSort ? -1 : rightSort ? 1 : 0;
     });
+
+    //Capture unique measures.
+    this.config.allMeasures = set(data.map(d => d[this.config.measure_col]))
+        .values()
+        .sort((a, b) => {
+            const leftSort = a < b,
+                rightSort = a > b;
+
+            if (this.config.measures && this.config.measures.length) {
+                const aPos = this.config.measures.indexOf(a),
+                    bPos = this.config.measures.indexOf(b),
+                    diff = aPos > -1 && bPos > -1 ? aPos - bPos : null;
+
+                return diff
+                    ? diff
+                    : aPos > -1 ? -1 : bPos > -1 ? 1 : leftSort ? -1 : rightSort ? 1 : 0;
+            } else return leftSort ? -1 : rightSort ? 1 : 0;
+        });
+    this.config.measures =
+        this.config.measures && this.config.measures.length
+            ? this.config.measures
+            : this.config.allMeasures;
+
     multiply(this, sortedData, this.config.measure_col);
 }
